+++ forwarder/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.1.0-dev
 * Initial release [#40](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/40)
+* Add stream id to Forwarder [#43](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/43)